refactor: update pywal palette generation to use type guards
Previously, I was checking the properties of the 'color' object using 'hasOwnProperty' to decide if the color
was a custom color (e.g. ffffff) or if it was a modified color (with a luminance modifier).
Doing it like that was very confusing and it didn't have full type checking and it was possible to pass in
properties that could not be used in conjuction with each other.

Using typesccripts typeguards, this has been reduced into a single if-statement, since typescript will automatically
calculate the inferred type: https://www.typescriptlang.org/docs/handbook/advanced-types.html#user-defined-type-guards.